Separate routes from app
